requirements.txt
Flask
requests
PyPDF2
python-dotenv
google-generativeai
Flask-OAuthlib